+++ dzVents/runtime/Utils.lua
@@ -8,12 +8,16 @@ local self = {
 	LOG_MODULE_EXEC_INFO = 2,
 	LOG_INFO = 3,
 	LOG_DEBUG = 4,
-	DZVERSION = '2.5.0', -- for Lua 5.3
+	DZVERSION = '2.5.4',
 }
 
 function math.pow(x, y)
 	self.log('Function math.pow(x, y) has been deprecated in Lua 5.3. Please consider changing code to x^y', self.LOG_FORCE)
-	return x^y 
+	return x^y
+end
+
+function self.setLogMarker(logMarker)
+	_G.logMarker = logMarker or _G.moduleLabel
 end
 
 function self.rightPad(str, len, char)
@@ -38,7 +42,7 @@ end
 function self.numDecimals(num, int, dec)
 	if int == nil then int = 99 end
 	if dec == nil then dec = 0 end
-	local fmt = '%' .. int .. '.' .. dec .. 'f' 
+	local fmt = '%' .. int .. '.' .. dec .. 'f'
 	return string.format(fmt,num)
 end
 
@@ -83,6 +87,11 @@ function self.round(x, n)
 		x = math.ceil(x - 0.5)
 	end
 	return x / n
+end
+
+function string.sMatch(text, match) -- add sanitized match function to string "library"
+	local sanitizedMatch = match:gsub("([%%%^%$%(%)%.%[%]%*%+%-%?])", "%%%1") -- escaping all 'magic' chars
+	return text:match(sanitizedMatch)
 end
 
 function self.toCelsius(f, relative)
@@ -150,6 +159,101 @@ function self.fromJSON(json, fallback)
 
 	self.log('Error parsing json to LUA table: ' .. results, self.LOG_ERROR)
 	return fallback
+
+end
+
+function self.fromBase64(codedString)  -- from http://lua-users.org/wiki/BaseSixtyFour
+	if type(codedString) ~= 'string' then
+		self.log('fromBase64: parm should be a string; you supplied a ' .. type(codedString), self.LOG_ERROR)
+		return nil
+	end
+	local b='ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/'
+	local data = string.gsub(codedString, '[^'.. b ..'=]', '')
+	return (data:gsub('.', function(x)
+		if (x == '=') then return '' end
+		local r, f = '',(b:find(x)-1)
+		for i = 6, 1, -1 do r = r .. (f%2^i-f%2^(i-1)>0 and '1' or '0') end
+		return r;
+	end):gsub('%d%d%d?%d?%d?%d?%d?%d?', function(x)
+		if (#x ~= 8) then return '' end
+		local c = 0
+		for i = 1, 8 do c = c + (x:sub(i, i) == '1' and 2^(8-i) or 0) end
+		return string.char(c)
+	end))
+end
+
+function self.toBase64(s) -- from http://lua-users.org/wiki/BaseSixtyFour
+	if type(s) == 'number' then s = tostring(s)
+	elseif type(s) ~= 'string' then
+		self.log('toBase64: parm should be a number or a string; you supplied a ' .. type(s), self.LOG_ERROR)
+		return nil
+	end
+	local bs =
+	{	[0] =
+				'A','B','C','D','E','F','G','H','I','J','K','L','M','N','O','P','Q','R','S','T','U','V','W','X','Y','Z',
+				'a','b','c','d','e','f','g','h','i','j','k','l','m','n','o','p','q','r','s','t','u','v','w','x','y','z',
+				'0','1','2','3','4','5','6','7','8','9',
+				'+','/'
+	}
+	local byte, rep = string.byte, string.rep
+	local pad = 2 - ((#s-1) % 3)
+	s = (s..rep('\0', pad)):gsub("...", function(cs)
+		local a, b, c = byte(cs, 1, 3)
+		return bs[a>>2] .. bs[(a&3)<<4|b>>4] .. bs[(b&15)<<2|c>>6] .. bs[c&63]
+	end)
+	return s:sub(1, #s-pad) .. rep('=', pad)
+end
+
+function self.fromXML(xml, fallback)
+
+	local parseXML = function(x)
+		local xmlParser = xml2Lua.parser(xmlHandler)
+		xmlParser:parse(x)
+		return xmlHandler.root
+	end
+
+	if xml == nil then
+		return fallback
+	end
+
+	if xml2Lua == nil then
+		xml2Lua = require('xml2lua')
+	end
+
+	if xmlHandler == nil then
+		xmlHandler = require("xmlhandler.tree")
+	end
+
+	ok, results = pcall(parseXML, xml)
+
+	if (ok) then
+		return results
+	end
+
+	self.log('Error parsing XML to LUA table: ' .. results, self.LOG_ERROR)
+	return fallback
+
+end
+
+function self.toXML(luaTable, header)
+	if header == nil then header = 'LuaTable' end
+
+	local toXML = function(luaTable, header)
+		return xmlParser.toXml(luaTable, header)
+	end
+
+	if (xmlParser == nil) then
+		xmlParser = require('xml2lua')
+	end
+
+	ok, results = pcall(toXML, luaTable, header)
+
+	if (ok) then
+		return results
+	end
+
+	self.log('Error converting LUA table to XML: ' .. results, self.LOG_ERROR)
+	return nil
 
 end
 
@@ -243,7 +347,7 @@ end
 local function loopGlobal(parm, baseType)
 	local res = 'id'
 	if type(parm) == 'number' then res = 'name' end
-	for i, item in ipairs(_G.domoticzData) do 
+	for i, item in ipairs(_G.domoticzData) do
 		if item.baseType == baseType and ( item.id == parm or item.name == parm ) then return item[res] end
 	end
 	return false